move fio fixtures to common conftest
