+++ tests/manage/monitoring/conftest.py
@@ -320,88 +320,6 @@ def measure_corrupt_pg(measurement_dir):
 #
 
 
-@pytest.fixture
-def fio_pvc_dict():
-    """
-    PVC template for fio workloads.
-    Note that all 'None' values needs to be defined before usage.
-    """
-    template = textwrap.dedent("""
-        kind: PersistentVolumeClaim
-        apiVersion: v1
-        metadata:
-          name: fio-target
-        spec:
-          storageClassName: None
-          accessModes: ["ReadWriteOnce"]
-          resources:
-            requests:
-              storage: None
-        """)
-    pvc_dict = yaml.safe_load(template)
-    return pvc_dict
-
-
-@pytest.fixture
-def fio_configmap_dict():
-    """
-    ConfigMap template for fio workloads.
-    Note that you need to add actual configuration to workload.fio file.
-    """
-    template = textwrap.dedent("""
-        kind: ConfigMap
-        apiVersion: v1
-        metadata:
-          name: fio-config
-        data:
-          workload.fio: |
-            # here comes workload configuration
-        """)
-    cm_dict = yaml.safe_load(template)
-    return cm_dict
-
-
-@pytest.fixture
-def fio_job_dict():
-    """
-    Job template for fio workloads.
-    """
-    template = textwrap.dedent("""
-        apiVersion: batch/v1
-        kind: Job
-        metadata:
-          name: fio
-        spec:
-          backoffLimit: 1
-          template:
-            metadata:
-              name: fio
-            spec:
-              containers:
-                - name: fio
-                  image: quay.io/johnstrunk/fs-performance:latest
-                  command:
-                    - "/usr/bin/fio"
-                    - "--output-format=json"
-                    - "/etc/fio/workload.fio"
-                  volumeMounts:
-                    - name: fio-target
-                      mountPath: /mnt/target
-                    - name: fio-config-volume
-                      mountPath: /etc/fio
-              restartPolicy: Never
-              volumes:
-                - name: fio-target
-                  persistentVolumeClaim:
-                    claimName: fio-target
-                - name: fio-config-volume
-                  configMap:
-                    name: fio-config
-        """)
-    job_dict = yaml.safe_load(template)
-    return job_dict
-
-
 def workload_fio_storageutilization(
     fixture_name,
     target_percentage,
